Replace old GetText method with new one
In 8.0 UIDropDownMenu_GetText stopped working - it always returns nil.
I can't find it mentioned anywhere so I looked for an alternative.  I
saw a number of addons using the `_G[dropDown:GetName() .. "Text"]`
method.  I don't fully understand it but I'm guessing that when you
create a UIDropDownMenuTemplate frame, the system creates a "Text" one
also.  So I'm using that.