Switched to using AceGUI to create dropdown
It was advised to me that using the built-in mechanism to create
dropdowns could taint the UI and instead to use the one in AceGUI.  So I
switched all three dropdowns to use that.  That actually cleaned up a
bunch because Ace handles the OnInit callback and setting the text.