Removed button modifiers on snooze
Turns out the Tutorial code only creates a single button which it then
modifies each time you turn the page.  So when I disable the button, it
disables it for all reminders.  And when I change the text, it's only
changed until you turn the page; going back to the previous page will
show "Snooze" instead of "Snoozed".