Set the BG textures of the frame
Turns out the UIPanelDialogTemplate frame has two textures - a
background for the Title section and one for the Dialog section.  So we
need to set those instead of the Background.  Setting the background
seems to be in a z-level fight with the textures.

Thanks to Kanegasi here - http://www.wowinterface.com/forums/showthread.php?p=329257#post329257

Fixes https://github.com/pcg79/Reminders-Addon/issues/11