Removed need for Lib:Ace3 addon
I thought I was doing this before but I was wrong.  Copied a bunch of
code from the Ace3 library and embedded it.

One thing - the Ace3 lib had a lot of what I think are duplicate files
(diff was the same and they ended in "(1)..." which is what MacOSX does
when you make a duplicate of a file).  So I deleted those.  Hope that
doesn't cause issues.